agrego Factura y demás errores

+++ src/servicios/Servicio.java
@@ -4,7 +4,6 @@ import containers.Container;
 
 public interface Servicio {
 	
-	
-	public int getPrecio(Container c);
+	public double getPrecio(Container c);
 
 }
